+++ src/main/java/com/hypherionmc/simplerpc/util/rpcavatar/StandardResponse.java
@@ -1,0 +1,7 @@
+package com.hypherionmc.simplerpc.util.rpcavatar;
+
+public class StandardResponse {
+    public boolean error;
+    public String message;
+    public String data;
+}